requirements.txt
pytest
numpy
opencv-python
Pillow